cleaner compile on FreeBSD. svn:NO TEST
git-svn-id: file:///svn/nsd/trunk@3257 a26ef69c-88ff-0310-839f-98b793d9c207

+++ server.c
@@ -782,7 +782,7 @@ server_reload(struct nsd *nsd, region_type* server_region, netio_type* netio,
 
 	/* if the parent has quit, we must quit too, poll the fd for cmds */
 	if(block_read(nsd, cmdsocket, &cmd, sizeof(cmd), 0) == sizeof(cmd)) {
-		DEBUG(DEBUG_IPC,1, (LOG_INFO, "reload: ipc command from main %d", cmd));
+		DEBUG(DEBUG_IPC,1, (LOG_INFO, "reload: ipc command from main %d", (int)cmd));
 		if(cmd == NSD_QUIT) {
 			DEBUG(DEBUG_IPC,1, (LOG_INFO, "reload: quit to follow nsd"));
 			send_children_quit(nsd);
@@ -820,7 +820,7 @@ server_reload(struct nsd *nsd, region_type* server_region, netio_type* netio,
 		log_msg(LOG_ERR, "reload: could not wait for parent to quit: %s",
 			strerror(errno));
 	}
-	DEBUG(DEBUG_IPC,1, (LOG_INFO, "reload: ipc reply main %d %d", ret, cmd));
+	DEBUG(DEBUG_IPC,1, (LOG_INFO, "reload: ipc reply main %d %d", ret, (int)cmd));
 	if(cmd == NSD_QUIT) {
 		/* small race condition possible here, parent got quit cmd. */
 		send_children_quit(nsd);
@@ -1160,7 +1160,7 @@ server_main(struct nsd *nsd)
 			nsd->mode = NSD_RUN;
 			break;
 		default:
-			log_msg(LOG_WARNING, "NSD main server mode invalid: %d", nsd->mode);
+			log_msg(LOG_WARNING, "NSD main server mode invalid: %d", (int)nsd->mode);
 			nsd->mode = NSD_RUN;
 			break;
 		}
@@ -1350,7 +1350,7 @@ server_child(struct nsd *nsd)
 			/* ignore here, quit */
 		} else {
 			log_msg(LOG_ERR, "mode bad value %d, back to service.",
-				mode);
+				(int)mode);
 			nsd->mode = NSD_RUN;
 		}
 	}
